Refactor HTTP helper stuff to dedicated folder
