+++ frontend/src/services/http/NetworkError.ts
